pyproject.toml
[tool.poetry]
name = "stract-api"
version = "1.0.0"
description = "api for stract"
authors = ["Everton Mutti <evertoncastro19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]
pre-commit = "^3.0.0"
pylint = "^2.17.2"
yapf = "^0.40.2"
ruff = "^0.1.6"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
